super basic rough swiping
