+++ DuckDuckGo/MainViewController.swift
@@ -216,129 +216,7 @@ class MainViewController: UIViewController {
         findInPageHeightLayoutConstraint = height
     }
     
-    class OmniBarCell: UICollectionViewCell {
-        
-        weak var omniBar: OmniBar? {
-            didSet {
-                subviews.forEach { $0.removeFromSuperview() }
-                if let omniBar {
-                    addSubview(omniBar)
-                    NSLayoutConstraint.activate([
-                        constrainView(omniBar, by: .leading),
-                        constrainView(omniBar, by: .trailing),
-                        constrainView(omniBar, by: .top),
-                        constrainView(omniBar, by: .bottom),
-                    ])
-                }
-            }
-        }
-        
-    }
-    
-    class NewTabCell: UICollectionViewCell {
-        static let identifier = "AddCell"
-
-        private let addButton: UIButton = {
-            let button = UIButton()
-            button.setTitle("+ Add", for: .normal)
-            button.backgroundColor = .systemBlue
-            button.layer.cornerRadius = 5
-            button.translatesAutoresizingMaskIntoConstraints = false
-            return button
-        }()
-
-        override init(frame: CGRect) {
-            super.init(frame: frame)
-            setupButton()
-        }
-
-        required init?(coder: NSCoder) {
-            fatalError("init(coder:) has not been implemented")
-        }
-
-        private func setupButton() {
-            contentView.addSubview(addButton)
-            addButton.addTarget(self, action: #selector(addButtonTapped), for: .touchUpInside)
-
-            // Auto Layout Constraints
-            let padding: CGFloat = 10
-            NSLayoutConstraint.activate([
-                addButton.topAnchor.constraint(equalTo: contentView.topAnchor, constant: padding),
-                addButton.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -padding),
-                addButton.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: padding),
-                addButton.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -padding)
-            ])
-        }
-
-        @objc private func addButtonTapped() {
-            // Handle the add button tap event
-            print("Add button tapped")
-        }
-    }
-
     var swipeTabsCoordinator: SwipeTabsCoordinator!
-    class SwipeTabsCoordinator: NSObject, UICollectionViewDataSource, UICollectionViewDelegate {
-        
-        weak var coordinator: MainViewCoordinator!
-        weak var tabsModel: TabsModel!
-        
-        init(coordinator: MainViewCoordinator) {
-            self.coordinator = coordinator
-            coordinator.navigationBarContainer.register(OmniBarCell.self, forCellWithReuseIdentifier: "omnibar")
-            coordinator.navigationBarContainer.register(NewTabCell.self, forCellWithReuseIdentifier: "newtab")
-        }
-        
-        func refresh(tabsModel: TabsModel, scrollToSelected: Bool = false) {
-            print("***", #function)
-            self.tabsModel = tabsModel
-            coordinator.navigationBarContainer.reloadData()
-        }
-        
-        func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-            tabsModel.count
-        }
-        
-        func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-            if indexPath.row > 0 {
-                return collectionView.dequeueReusableCell(withReuseIdentifier: "newtab", for: indexPath)
-            } else {
-                let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "omnibar", for: indexPath) as! OmniBarCell
-                cell.omniBar = coordinator.omniBar
-                return cell
-            }
-        }
-        
-        var startOffsetX: CGFloat = 0.0
-        var startingIndexPath: IndexPath?
-        func scrollViewWillBeginDragging(_ scrollView: UIScrollView) {
-            startOffsetX = scrollView.contentOffset.x
-            startingIndexPath = coordinator.navigationBarContainer.indexPathsForVisibleItems[0]
-        }
-        
-        var targetIndexPath: IndexPath?
-        func scrollViewDidScroll(_ scrollView: UIScrollView) {
-            guard let startingIndexPath else { return }
-            let distance = scrollView.contentOffset.x - startOffsetX
-            print("***", #function, startingIndexPath, distance)
-            if abs(distance) > coordinator.superview.frame.width * 0.3 {
-                var targetIndexPath = startingIndexPath
-                if distance < 0 {
-                    targetIndexPath.row -= 1
-                } else {
-                    targetIndexPath.row += 1
-                }
-                self.targetIndexPath = targetIndexPath
-            } else {
-                targetIndexPath = nil
-            }
-        }
-
-        func scrollViewDidEndDragging(_ scrollView: UIScrollView, willDecelerate decelerate: Bool) {
-            guard let targetIndexPath else { return }
-            print("***", #function, targetIndexPath)
-            coordinator.navigationBarContainer.scrollToItem(at: targetIndexPath, at: .centeredHorizontally, animated: true)
-        }
-    }
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -350,7 +228,8 @@ class MainViewController: UIViewController {
         viewCoordinator.toolbarForwardButton.action = #selector(onForwardPressed)
         viewCoordinator.toolbarFireButton.action = #selector(onFirePressed)
 
-        swipeTabsCoordinator = SwipeTabsCoordinator(coordinator: viewCoordinator)
+        swipeTabsCoordinator = SwipeTabsCoordinator(coordinator: viewCoordinator, tabPreviewsSource: previewsSource, selectTab: select)
+                                                    
         viewCoordinator.navigationBarContainer.delegate = swipeTabsCoordinator
         viewCoordinator.navigationBarContainer.dataSource = swipeTabsCoordinator
         
@@ -1000,7 +879,7 @@ class MainViewController: UIViewController {
             refreshControls()
         }
         tabsBarController?.refresh(tabsModel: tabManager.model, scrollToSelected: true)
-        tabsBarController?.refresh(tabsModel: tabManager.model, scrollToSelected: true)
+        swipeTabsCoordinator?.refresh(tabsModel: tabManager.model, scrollToSelected: true)
         if DaxDialogs.shared.shouldShowFireButtonPulse {
             showFireButtonPulse()
         }
